backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.23.1",
    "azure-search-documents>=11.5.3",
    "fastapi[standard]>=0.116.1",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
    "semantic-kernel>=1.32.2",
]
